requirements.txt
requests_html
psycopg2-binary
python-dotenv
